pyproject.toml
[tool.poetry]
name = "fastapi-graphql-masonite-orm"
version = "0.1.0"
description = ""
authors = ["myg0t <mjtaylorroberts@gmail.com>"]

[tool.black]
line-length = 99 # override black's default line-length
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.76.0"
masonite-orm = "^2.9.2"
graphene = "^3.0"
starlette-graphene3 = "^0.5.1"
python-dotenv = "^0.20.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
typer = "^0.4.1"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
